remzi: separate test_common from test_memory_alloc to use it in other unit tests
