+++ unit_tests/common/test_debug.hpp
@@ -1,0 +1,41 @@
+/*
+ * Copyright (C) 2008 Search Solution Corporation. All rights reserved by Search Solution.
+ *
+ *   This program is free software; you can redistribute it and/or modify
+ *   it under the terms of the GNU General Public License as published by
+ *   the Free Software Foundation; either version 2 of the License, or
+ *   (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program; if not, write to the Free Software
+ *  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
+ *
+ */
+ 
+/*
+ * test_debug.hpp - debuggers for unit tests
+ */
+
+#ifndef _TEST_DEBUG_HPP_
+#define _TEST_DEBUG_HPP_
+
+namespace test_common {
+
+// crash on release too
+inline void
+custom_assert (bool cond)
+{
+  if (!cond)
+    {
+      abort ();
+    }
+}
+
+} // namespace test_common
+
+#endif _TEST_DEBUG_HPP_
